Align agenda to right side of screen
Not the same as `text-align: right`. Want the text left aligned, but to prefer the right side of the screen. Ignore the width of the hidden text underneath it. So it's not floating over the middle of the broadcast.

Causes the collapse toggle to fly in from way above. But that's better than no animation, e.g. via `display: none`.